# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Whisper server",
        "description": "This is Whisper messaging server.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Creates new account.",
                "operationId": "RegisterUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                },
                                "example": {
                                    "error": "..."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                },
                                "example": {
                                    "error": "..."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify": {
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verifies account.",
                "operationId": "VerifyUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyUserInput"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                },
                                "example": {
                                    "error": "..."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                },
                                "example": {
                                    "error": "..."
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterUserInput": {
                "type": "object",
                "properties": {
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "str0ng_p455w0rd",
                        "maxLength": 64,
                        "minLength": 8
                    },
                    "username": {
                        "type": "string",
                        "example": "ex4ample",
                        "maxLength": 32,
                        "minLength": 2
                    }
                },
                "required": [
                    "mail",
                    "password",
                    "username"
                ]
            },
            "VerifyUserInput": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "tQrpaCeDwBD71217HXBG9Y35rg1ECT78",
                        "maxLength": 32,
                        "minLength": 32
                    }
                },
                "required": [
                    "code"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Put **_ONLY_** your JWT Bearer token on textbox below!"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "User authentication."
        }
    ]
}